Implemented Trac #159.
This adds support for intelligent import from the clipboard (as the paste function).  The old-style textbox import is still around for the time being; it will be killed off by a future change that makes the import dialog into a standard File->Open dialog box instead of the current hacky-box.