Added new supported items in itemRemove item class.
Added the ability to remove charges and projected items using the
itemRemove built in context menu.

In this commit I fixed a minor bug in the projected view that caused the
right clicked item not to be SHOWN as selected until after the context
menu had been spawned and closed.

I also removed all parent references from the ContextMenu class and sub
classes. I am instead using the fit service to do all the work of
removing the items in the itemRemove class.