Integrated Hacking Modules Give Double Bonus Incorrectl
Because integrated modules are a single module with both Data and Relic
analizer built into one, ship hulls were giving a bonus twice over.
Changed it from adding the bonus for each individually (which worked
fine when it was broken out) to simply adding it if either Hacking or
Archaeology skill is present.