Applied Mining Crystal Bonus to modules
Previously the mining crystal bonus was not applied to modules, so
adding a crystal didn't change the amount mined at all.  This now
applies.  Left the old application in place, so it still applies to a
nonsense `specialtyMiningAmount` stat if you want to see that seperated
out for some weird reason.