Fix cargo items price increase on refit
The price attribute of an item is linked to its price in the database,
through SQLAlchemy and the ORM map. Updating a price attribute therefore
updates the database. This patch changes the calculation of the cost of
amounts of items, so that the price in the database no longer occurs.

Fixes #1759